Backend/requirements.txt
asgiref==3.7.2
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.7.0
Django==4.2.3
django-cors-headers==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fonttools==4.47.2
html5lib==1.1
idna==3.6
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.0.0
psycopg==3.1.17
pycparser==2.21
pydyf==0.8.0
Pygments==2.17.2
PyJWT==2.8.0
pyphen==0.14.0
python-dotenv==1.0.0
pytz==2023.3
redgreenunittest==0.1.1
reportlab==4.0.9
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
weasyprint==60.2
webencodings==0.5.1
zopfli==0.2.3
